code/requirements.txt
numpy>=2.3
pandas>=2.3
matplotlib>=3.10
statsmodels>=0.14
openpyxl>=3.1
pyyaml>=6.0
scikit-learn
